update dashboard analysis layout
